refactor storages
