+++ src/storages/InstitutesStorage.js
@@ -1,0 +1,15 @@
+import axios from "axios";
+import { defineStore } from "pinia";
+export const useInstitutesStorage = defineStore("InstitutesStorage", {
+  state: () => ({
+    Institutes: [],
+  }),
+  actions: {
+    async getInstitutesFromServer() {
+      axios.get("/api/out/legacy/faculty")
+      .then((response) => {
+          this.Insts = response.data;
+        });
+    },
+  }
+})
